+++ train/strategies/sft.py
@@ -44,7 +44,7 @@ class SFTStrategy(AbstractTrainingStrategy):
             zero_loss: torch.Tensor = (shift_logits.sum() * 0.0).float()
             return zero_loss
         return cross_entropy(
-            shift_logits.view(-1, shift_logits.size(-1)),
+            shift_logits.float().view(-1, shift_logits.size(-1)),
             shift_labels.view(-1),
             ignore_index=-100,
         )
